+++ packages/liveblocks-core/src/client.ts
@@ -1,7 +1,7 @@
 import { deprecateIf } from "./lib/deprecation";
-import type { InternalRoom, Polyfills } from "./room";
+import type { InternalRoom, Polyfills, RoomInitializers } from "./room";
 import { createRoom } from "./room";
-import type { Authentication, RoomInitializers } from "./types";
+import type { Authentication } from "./types";
 import type { BaseUserMeta } from "./types/BaseUserMeta";
 import type { Json, JsonObject } from "./types/Json";
 import type { LsonObject } from "./types/Lson";
